+++ 1-testAnimation/src/Data/melodies.js
@@ -1,0 +1,83 @@
+// Collection of recognizable melodies that work well with irregular bouncing
+export const melodies = {
+    // Happy Birthday - Simple, iconic intervals
+    happyBirthday: [
+        'C4', 'C4', 'D4', 'C4', 'F4', 'E4',
+        'C4', 'C4', 'D4', 'C4', 'G4', 'F4',
+        'C4', 'C4', 'C5', 'A4', 'F4', 'E4', 'D4',
+        'Bb4', 'Bb4', 'A4', 'F4', 'G4', 'F4'
+    ],
+
+    // Twinkle Twinkle - Perfect for inconsistent timing
+    twinkleTwinkle: [
+        'C4', 'C4', 'G4', 'G4', 'A4', 'A4', 'G4',
+        'F4', 'F4', 'E4', 'E4', 'D4', 'D4', 'C4',
+        'G4', 'G4', 'F4', 'F4', 'E4', 'E4', 'D4',
+        'G4', 'G4', 'F4', 'F4', 'E4', 'E4', 'D4',
+        'C4', 'C4', 'G4', 'G4', 'A4', 'A4', 'G4',
+        'F4', 'F4', 'E4', 'E4', 'D4', 'D4', 'C4'
+    ],
+
+    // Mary Had a Little Lamb - Simple and memorable
+    maryLamb: [
+        'E4', 'D4', 'C4', 'D4', 'E4', 'E4', 'E4',
+        'D4', 'D4', 'D4', 'E4', 'G4', 'G4',
+        'E4', 'D4', 'C4', 'D4', 'E4', 'E4', 'E4',
+        'D4', 'D4', 'E4', 'D4', 'C4'
+    ],
+
+    // Ode to Joy - Beethoven's iconic melody
+    odeToJoy: [
+        'E4', 'E4', 'F4', 'G4', 'G4', 'F4', 'E4', 'D4',
+        'C4', 'C4', 'D4', 'E4', 'E4', 'D4', 'D4',
+        'E4', 'E4', 'F4', 'G4', 'G4', 'F4', 'E4', 'D4',
+        'C4', 'C4', 'D4', 'E4', 'D4', 'C4', 'C4'
+    ],
+
+    // Jingle Bells - Instantly recognizable
+    jingleBells: [
+        'E4', 'E4', 'E4', 'E4', 'E4', 'E4',
+        'E4', 'G4', 'C4', 'D4', 'E4',
+        'F4', 'F4', 'F4', 'F4', 'F4', 'E4', 'E4', 'E4',
+        'E4', 'D4', 'D4', 'E4', 'D4', 'G4'
+    ],
+
+    // Old MacDonald - Fun and bouncy
+    oldMacDonald: [
+        'C4', 'C4', 'C4', 'G4', 'A4', 'A4', 'G4',
+        'E4', 'E4', 'D4', 'D4', 'C4',
+        'G4', 'G4', 'F4', 'F4', 'E4', 'E4', 'D4',
+        'C4', 'C4', 'C4', 'G4', 'A4', 'A4', 'G4',
+        'E4', 'E4', 'D4', 'D4', 'C4'
+    ],
+
+    // Imperial March (Star Wars) - Dramatic and recognizable
+    imperialMarch: [
+        'G4', 'G4', 'G4', 'Eb4', 'Bb4', 'G4', 'Eb4', 'Bb4', 'G4',
+        'D5', 'D5', 'D5', 'Eb5', 'Bb4', 'Gb4', 'Eb4', 'Bb4', 'G4',
+        'G5', 'G4', 'G4', 'G5', 'Gb5', 'F5', 'E5', 'Eb5', 'E5'
+    ],
+
+    // Für Elise opening - Classical and distinctive
+    furElise: [
+        'E5', 'Eb5', 'E5', 'Eb5', 'E5', 'B4', 'D5', 'C5', 'A4',
+        'C4', 'E4', 'A4', 'B4',
+        'E4', 'Ab4', 'B4', 'C5',
+        'E4', 'E5', 'Eb5', 'E5', 'Eb5', 'E5', 'B4', 'D5', 'C5', 'A4'
+    ],
+
+    // Nokia Ringtone - Instantly recognizable
+    nokiaRingtone: [
+        'E5', 'D5', 'Gb4', 'Ab4',
+        'C5', 'Bb4', 'D4', 'E4',
+        'B4', 'A4', 'Db4', 'E4',
+        'A4'
+    ],
+
+    // Super Mario Bros Theme - Gaming classic
+    marioTheme: [
+        'E5', 'E5', 'E5', 'C5', 'E5', 'G5', 'G4',
+        'C5', 'G4', 'E4', 'A4', 'B4', 'Bb4', 'A4',
+        'G4', 'E5', 'G5', 'A5', 'F5', 'G5', 'E5', 'C5', 'D5', 'B4'
+    ]
+};
