+++ 1-testAnimation/src/Data/noteFrequencies.js
@@ -1,0 +1,27 @@
+// Musical note frequencies (in Hz) - Extended range
+export const noteFrequencies = {
+    'C4': 261.63,
+    'Db4': 277.18,
+    'D4': 293.66,
+    'Eb4': 311.13,
+    'E4': 329.63,
+    'F4': 349.23,
+    'Gb4': 369.99,
+    'G4': 392.00,
+    'Ab4': 415.30,
+    'A4': 440.00,
+    'Bb4': 466.16,
+    'B4': 493.88,
+    'C5': 523.25,
+    'Db5': 554.37,
+    'D5': 587.33,
+    'Eb5': 622.25,
+    'E5': 659.25,
+    'F5': 698.46,
+    'Gb5': 739.99,
+    'G5': 783.99,
+    'Ab5': 830.61,
+    'A5': 880.00,
+    'Bb5': 932.33,
+    'B5': 987.77
+};
