+++ 1-testAnimation/src/Data/screen.js
@@ -1,0 +1,19 @@
+export const mouse = {
+	x: innerWidth / 2,
+	y: innerHeight / 2
+}
+
+// Camera system
+export const camera = {
+    x: 0,
+    y: 0,
+    targetX: 0,
+    targetY: 0,
+    smoothing: 0.08, 
+    shakeX: 0,
+    shakeY: 0,
+    shakeIntensity: 0,
+    shakeDecay: 0.9
+};
+
+export const colors = ['#ff0000', '#00ff00', '#0000ff', '#ff00ff'];
